change exceptions matching http error codes add jcasbin as dependency mark hasAuth deprecated since this project is using jcasbin update dependency refactor exception handler refactor mapper repository refactor rate limit service refactor token acquire add best answer support fix wrong annotation issue of section

+++ src/main/kotlin/com/young_zy/forum/service/MessageService.kt
@@ -1,10 +1,13 @@
 package com.young_zy.forum.service
 
+import com.young_zy.forum.common.exception.ForbiddenException
 import com.young_zy.forum.model.message.DetailedMessage
 import com.young_zy.forum.model.message.MessageEntity
 import com.young_zy.forum.repo.MessageNativeRepository
-import com.young_zy.forum.service.exception.NotFoundException
+import com.young_zy.forum.common.exception.NotFoundException
 import kotlinx.coroutines.flow.toList
+import kotlinx.coroutines.reactive.awaitSingleOrNull
+import org.casbin.jcasbin.main.Enforcer
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -29,16 +32,22 @@ class MessageService {
     @Autowired
     private lateinit var authService: AuthService
 
-    suspend fun postMessage(token: String, to: Long, messageText: String) {
+    @Autowired
+    private lateinit var enforcer: Enforcer
+
+    suspend fun postMessage(to: Long, messageText: String) {
         logger.info("post message start")
-        val tokenObj = loginService.getToken(token)
-        authService.hasAuth(tokenObj, AuthConfig(AuthLevel.USER))
+        val tokenObj = loginService.getToken()
+//        authService.hasAuth(tokenObj, AuthConfig(AuthLevel.USER))
+        if (!enforcer.enforce(tokenObj, "", "postMessage")) {
+            throw ForbiddenException("permission denied")
+        }
         transactionalOperator.executeAndAwait {
             logger.info("post message transaction start")
             val message = MessageEntity(
-                    from = tokenObj!!.uid,
-                    to = to,
-                    messageText = messageText
+                from = tokenObj!!.uid,
+                to = to,
+                messageText = messageText
             )
             messageNativeRepository.insertMessage(message)
             logger.info("post message transaction end")
@@ -46,38 +55,49 @@ class MessageService {
         logger.info("post message end")
     }
 
-    suspend fun deleteMessage(token: String, messageId: Long){
-        val tokenObj = loginService.getToken(token)
+    suspend fun deleteMessage(messageId: Long) {
+        val tokenObj = loginService.getToken()
         transactionalOperator.executeAndAwait {
-            val message = messageNativeRepository.getMessageById(messageId) ?: throw NotFoundException("message with messageId $messageId not found")
-            authService.hasAuth(tokenObj, AuthConfig(
-                    authLevel =  AuthLevel.USER,
-                    allowOnlyAuthor = true,
-                    authorUid = message.to
-            ))
+            val message = messageNativeRepository.getMessageById(messageId).awaitSingleOrNull()
+                ?: throw NotFoundException("message with messageId $messageId not found")
+//            authService.hasAuth(tokenObj, AuthConfig(
+//                    authLevel =  AuthLevel.USER,
+//                    allowOnlyAuthor = true,
+//                    authorUid = message.to
+//            ))
+            if (!enforcer.enforce(tokenObj, message, "deleteMessage")) {
+                throw ForbiddenException("permission denied")
+            }
             messageNativeRepository.deleteMessage(messageId)
         }
     }
 
-    suspend fun setReadState(token: String, messageId: Long, readState: Boolean){
-        val tokenObj = loginService.getToken(token)
+    suspend fun setReadState(messageId: Long, readState: Boolean) {
+        val tokenObj = loginService.getToken()
         transactionalOperator.executeAndAwait {
-            val message = messageNativeRepository.getMessageById(messageId) ?: throw NotFoundException("message with messageId $messageId not found")
-            authService.hasAuth(tokenObj, AuthConfig(
-                    authLevel =  AuthLevel.USER,
-                    allowOnlyAuthor = true,
-                    authorUid = message.to
-            ))
+            val message = messageNativeRepository.getMessageById(messageId).awaitSingleOrNull()
+                ?: throw NotFoundException("message with messageId $messageId not found")
+//            authService.hasAuth(tokenObj, AuthConfig(
+//                    authLevel =  AuthLevel.USER,
+//                    allowOnlyAuthor = true,
+//                    authorUid = message.to
+//            ))
+            if (!enforcer.enforce(tokenObj, message, "deleteMessage")) {
+                throw ForbiddenException("permission denied")
+            }
             message.unread = readState
             messageNativeRepository.updateMessage(message)
         }
     }
 
-    suspend fun getAllMessage(token: String, page: Long, size: Long): List<DetailedMessage> {
-        val tokenObj = loginService.getToken(token)
-        authService.hasAuth(tokenObj, AuthConfig(
-                authLevel =  AuthLevel.BLOCKED_USER
-        ))
+    suspend fun getAllMessage(page: Long, size: Long): List<DetailedMessage> {
+        val tokenObj = loginService.getToken()
+//        authService.hasAuth(tokenObj, AuthConfig(
+//                authLevel =  AuthLevel.BLOCKED_USER
+//        ))
+        if (!enforcer.enforce(tokenObj, "", "getAllMessage")) {
+            throw ForbiddenException("permission denied")
+        }
         return messageNativeRepository.getAllByUid(tokenObj!!.uid, page, size).toList()
     }
